+++ package.json
@@ -61,6 +61,10 @@
         "lines": 80
       }
     },
+    "testMatch": [
+      "<rootDir>/src/**/__tests__/**/*.tsx",
+      "<rootDir>/src/**/?(*.)+(spec|test).tsx"
+    ],
     "coveragePathIgnorePatterns": [
       "src/server",
       "src/index.tsx",
