+++ src/__tests__/Sort.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen, waitFor } from '@testing-library/react'
-import Sort from './Sort'
+import Sort from '../Sort'
 import userEvent from '@testing-library/user-event'
 import { act } from 'react-dom/test-utils'
